refactors code and extracts helper function for data handling
