+++ src/model/censusdata/services/geodataApiDataService.js
@@ -1,4 +1,5 @@
 import { csvParse } from "d3-dsv";
+import {writeDataToMapObj} from "../../utils"
 
 const baseURL = "https://5laefo1cxd.execute-api.eu-central-1.amazonaws.com/dev/hello/skinny";
 
@@ -30,17 +31,7 @@ export default class GeodataApiDataService {
     }
     const response = await fetch(url);
     const string = await response.text();
-    let data = new Map();
-    csvParse(string, (row, i, cols) => {
-      let geoDataObject = {};
-      cols.forEach((col, i) => {
-        if (i > 0) {
-          geoDataObject[col] = +row[col];
-        }
-      });
-      data.set(row.geography_code, geoDataObject);
-    });
-    return data;
+    return writeDataToMapObj(string);
   }
 
   async fetchSelectedDataForGeographies(geoCodes, catCodes) {
@@ -49,17 +40,7 @@ export default class GeodataApiDataService {
     const url = `${baseURL}?cols=geography_code,${catCodesString}&rows=${geoCodesString}`;
     const response = await fetch(url);
     const string = await response.text();
-    let data = new Map();
-    csvParse(string, (row, i, cols) => {
-      let geoDataObject = {};
-      cols.forEach((col, i) => {
-        if (i > 0) {
-          geoDataObject[col] = +row[col];
-        }
-      });
-      data.set(row.geography_code, geoDataObject);
-    });
-    return data;
+    return writeDataToMapObj(string);
   }
 
   async fetchSelectedDataForBoundingBox(geoType, catCodes, bBox) {
@@ -68,20 +49,6 @@ export default class GeodataApiDataService {
     const url = `${baseURL}?bbox=${bBox.neCorner.lng},${bBox.neCorner.lat},${bBox.swCorner.lng},${bBox.swCorner.lat}&cols=geography_code,${catCodesString}&geotype=${geoTypesString}`;
     const response = await fetch(url);
     const string = await response.text();
-    let data = new Map();
-    csvParse(string, (row, i, cols) => {
-      let geoDataObject = {};
-      cols.forEach((col, i) => {
-        if (i > 0) {
-          geoDataObject[col] = +row[col];
-        }
-      });
-      data.set(row.geography_code, geoDataObject);
-    });
-    return data;
-  }
-
-  async fetchCensusTableStructure() {
-    return simpleTopicTableCategoryData;
+    return writeDataToMapObj(string);
   }
 }
