+++ src/model/utils.js
@@ -1,3 +1,5 @@
+import { csvParse } from "d3-dsv";
+
 export function getLegendSection(value, breakpoints) {
   for (let i = 1; i < breakpoints.length; i++) {
     if (value <= breakpoints[i]) {
@@ -5,4 +7,18 @@ export function getLegendSection(value, breakpoints) {
     }
   }
   return breakpoints.length;
+}
+
+export function writeDataToMapObj(responseStr){
+  let data = new Map();
+    csvParse(responseStr, (row, i, cols) => {
+      let geoDataObject = {};
+      cols.forEach((col, i) => {
+        if (i > 0) {
+          geoDataObject[col] = +row[col];
+        }
+      });
+      data.set(row.geography_code, geoDataObject);
+    });
+  return data
 }
